backend/requirements.txt
altair==5.2.0
auditok==0.2.0
extra_streamlit_components==0.1.60
ffmpeg-python==0.2.0
graphviz==0.20.1
tensorflow==2.15.0
keras==2.15.0
keyboard==0.13.5
numpy==1.26.3
openai==1.10.0
pandas==2.2.0
pdfminer.six==20221105
Pillow==10.2.0
plotly==5.18.0
pyaudio==0.2.14
PyDrive==1.3.1
pydub==0.25.1
pymongo==4.6.0
regex==2023.12.25
Requests==2.31.0
sentence_transformers==2.2.2
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
st_pages==0.4.5
streamlit==1.28.0
streamlit_extras==0.3.6
streamlit_scrollable_textbox==0.0.3
torch==2.1.2
whisper_timestamped==1.14.4
opencv-python==4.9.0.80
